Grading pipeline: Restart logic
Since these jobs take days to run, I am implemened a restart logic.
Given an unfinished paragraphs-file, the grading pipeline will copy each
query's annotations until the file is exhausted (or restart-from-query
is seen), then runs the pipeline on remaining queries.

it is assumed that both the input paragraphs-file, and the unfinished
paragraphs-file list queries in the same order. A check raises an
exception if this is not the case.