adjust the test bank code for new pydantic
pydantic changed how it is handling the parsing of generics.
This means that the object type needs to be known at parsing time, for
fields of subtypes to be properly parsed.

Hence, we introduce a factory method for parsing nuggets vs questions in
homogeneous types.

I am not super happy with the solution, but it works now.