exam_embed_train: avoid long debug message on missing
Since a label/grade of -1 indicates a missing value,
when we convert this into a one-hot, we have to replace it with a value
in the dense range (e.g default_val = 0)

We want to print a warning on unexpected values, but this message should
be suppressed on -1. However the default_value should be set either way.